+++ lib/chef-acceptance/cli.rb
@@ -1,6 +1,7 @@
 require "thor"
 require "chef-acceptance/version"
 require "chef-acceptance/application"
+require "chef-acceptance/options"
 require "chef-acceptance/test_suite"
 require "chef-acceptance/acceptance_cookbook"
 require "chef-acceptance/executable_helper"
@@ -10,20 +11,39 @@ module ChefAcceptance
     package_name "chef-acceptance"
 
     #
+    # Define shared options
+    #
+    option_timeout = [
+      :timeout,
+      type: :numeric,
+      desc: "Override default chef-client timeout. (Default: #{ChefAcceptance::Options::DEFAULT_TIMEOUT})",
+    ]
+
+    option_audit_mode = [
+      :audit_mode,
+      type: :boolean,
+      desc: "Enable or disable audit_mode (Default: true)",
+    ]
+
+    #
     # Create core acceptance commands
     #
     AcceptanceCookbook::CORE_ACCEPTANCE_RECIPES.each do |command|
       desc "#{command} TEST_SUITE_REGEX", "Run #{command}"
+      option(*option_timeout)
+      option(*option_audit_mode)
       define_method(command) do |test_suite_regex = ".*"|
-        app = Application.new()
+        app = Application.new(options)
         app.run(test_suite_regex, command)
       end
     end
 
-    desc "test TEST_SUITE_REGEX [--force-destroy]", "Run provision, verify and destroy"
+    desc "test TEST_SUITE_REGEX", "Run provision, verify and destroy"
     option :force_destroy,
            type: :boolean,
            desc: "Force destroy phase after any run"
+    option(*option_timeout)
+    option(*option_audit_mode)
     def test(test_suite_regex = ".*")
       app = Application.new(options)
       app.run(test_suite_regex, "test")